+++ CHANGELOG.md
@@ -56,6 +56,7 @@
 - Fixed spelling of `DIRECTIONS` in documentation for routing tool.
 - Tool CoordToolkit: Labeling of longitude and latitude corrected if long/lat is selected as projection system.
 - Fixed issue #672: layerAttribution not working correctly if layer is not initially visible
+- Issue #673: LayerSlider: The progress bar is now displayed correctly with more than 10 configured layers.
 
 ---
 ## v2.17.1 - 2022-01-011